fix(rabbitmq): accept non-int prefetch setting overrides
The prefetch override unboxed the value as int before conversion, so any
boxed long / ushort / string from a configuration provider threw
InvalidCastException. Convert.ToUInt16 already handles every IConvertible,
so drop the explicit cast and let it negotiate the numeric/string input.

Refs: M23

+++ src/ServiceConnect.Client.RabbitMQ/RabbitMqConsumerHost.cs
@@ -86,7 +86,7 @@ internal sealed class RabbitMqConsumerHost : IAsyncDisposable
         _errorsDisabled = queueConfiguration.DisableErrors;
         _autoDelete = settings.TryGetValue(RabbitMQSettingKeys.AutoDelete, out var autoDeleteVal) && (bool)autoDeleteVal;
         _prefetchCount = settings.TryGetValue(RabbitMQSettingKeys.PrefetchCount, out var prefetchVal)
-            ? Convert.ToUInt16((int)prefetchVal)
+            ? Convert.ToUInt16(prefetchVal)
             : transportConfiguration.PrefetchCount;
         _disablePrefetch = settings.TryGetValue(RabbitMQSettingKeys.DisablePrefetch, out var disablePrefetchVal) && (bool)disablePrefetchVal;
         _queueArguments = settings.TryGetValue(RabbitMQSettingKeys.Arguments, out var argsVal)
